Refactor `amisafe` into `isnodesafe`
